Complete Sharpe ratio in Portfolio class
Modified a little bit of other classes with overloaded methods, doesn't affect normal use

+++ src/main/java/com/ibm/security/appscan/altoromutual/model/Position.java
@@ -93,7 +93,7 @@ public class Position {
         for (double currentNum : table) {
             total += currentNum;
         }
-        return total / table.size();
+        return total / (double)table.size();
     }
 
     /**
